Initial support for stack mixin
